Remove IView and rename view packages
