--- src/main/java/com/bayou/views/IView.java
@@ -1,10 +1,0 @@
-package com.bayou.views;
-
-/**
- * Created by joshuaeaton on 1/31/17.
- */
-public interface IView {
-    Long getId();
-
-    void setId(Long id);
-}
