+++ src/main/java/com/bayou/views/UnverifiedUserView.java
@@ -1,26 +1,13 @@
-package com.bayou.views.impl;
-
-import com.bayou.views.IView;
+package com.bayou.views;
 
 /**
  * Created by rachelguillory 2/16/2017.
  */
-public class UnverifiedUserView implements IView {
-    private Long id;
+public class UnverifiedUserView extends BaseEntityView {
     private String passwordHash;
     private String passwordSalt;
     private String email;
     private Integer verificationCode;
-
-    @Override
-    public Long getId() {
-        return id;
-    }
-
-    @Override
-    public void setId(Long id) {
-        this.id = id;
-    }
 
     public String getPasswordHash() {
         return passwordHash;
@@ -58,10 +45,10 @@ public class UnverifiedUserView implements IView {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
+        if (!super.equals(o)) return false;
 
         UnverifiedUserView that = (UnverifiedUserView) o;
 
-        if (id != null ? !id.equals(that.id) : that.id != null) return false;
         if (passwordHash != null ? !passwordHash.equals(that.passwordHash) : that.passwordHash != null) return false;
         if (passwordSalt != null ? !passwordSalt.equals(that.passwordSalt) : that.passwordSalt != null) return false;
         if (email != null ? !email.equals(that.email) : that.email != null) return false;
@@ -70,7 +57,7 @@ public class UnverifiedUserView implements IView {
 
     @Override
     public int hashCode() {
-        int result = id != null ? id.hashCode() : 0;
+        int result = super.hashCode();
         result = 31 * result + (passwordHash != null ? passwordHash.hashCode() : 0);
         result = 31 * result + (passwordSalt != null ? passwordSalt.hashCode() : 0);
         result = 31 * result + (email != null ? email.hashCode() : 0);
